+++ conf_t/__init__.py
@@ -1,2 +1,2 @@
 # Conf T package initialization
-__version__ = "0.3.1"
+__version__ = "0.3.2"